corex(backend): CUDA backend Iluvatar Corex 4.4 compatibility
This is the bulk of the corex porting effort. Changes touch the
CUDA backend in five areas:

1. linear_system/
   - linear_pcg_corex.{cu.inc,h} + linear_fused_pcg_corex.{cu,h}
     replace the textbook PCG inner loop on corex. Critical fix:
     the convergence test is changed from the unsafe |r.z| <= tol
     to the textbook ||r|| <= max(rate*||b||, tol). The old test
     was masking divergence on corex (some scenes "converged" with
     residual still ~1e3) which produced apparent speedups but
     incorrect simulations on this platform.
   - linear_pcg.{cu,h} + linear_fused_pcg.{cu,h}: dispatch into the
     corex sidecar when UIPC_MUDA_USE_COREX is on; behavior on NVIDIA
     CUDA is unchanged.

2. affine_body/
   - abd_jacobi_matrix_corex.{cu,h}: corex-friendly variant of the
     ABD Jacobi block ops (avoids a few muda primitives the corex
     toolchain mis-codegens).
   - inter_affine_body_constitution.{cu,h}: cache the manager pointer
     as a SimSystemSlot member so the new corex for_each() template
     can route through the FilteredInfo without re-resolving the
     manager each call.
   - abd_diag_preconditioner / abd_linear_subsystem /
     abd_line_search_reporter / abd_active_set_reporter /
     abd_tolerance_checker / affine_body_dynamics: minor
     refactoring + corex sidecar inclusion guards.

3. collision_detection/
   - linear_bvh_corex.h, stackless_bvh_corex.h,
     info_stackless_bvh_corex.h, info_stackless_bvh_v0_corex.h:
     corex-specific BVH traversal (the original recursion pattern
     hits Iluvatar codegen bugs / shared-mem limits).
   - corresponding non-corex headers gain the include-sidecar
     dispatch.

4. finite_element/
   - mas_preconditioner_engine_corex.{cu,h}: corex variant of the
     MAS preconditioner (still slower than NVIDIA path -- see
     porting-build-run-guide.md for the SpMV-count regression).
   - fem_mas_preconditioner.cu: gain a non-corex
     engine.set_preconditioner() overload so the same call site
     compiles on both NVIDIA CUDA and Corex.
   - fem_utils.cu: under non-corex builds, keep the inline FEM
     deformation kernels (Ds / Dm_inv / F / dFdx) here so corex
     can pull them from fem_utils_corex.h (no-RDC corex codegen).
   - fem_utils_corex.h, matrix_utils_corex.h: float32-safe
     reductions and small-matrix kernels.

5. active_set_system / contact_system / type_define
   - global_active_set_manager: corex-safe atomic flush path.
   - contact_coeff: float32 tolerance widening.
   - type_define.h: extra Corex-only typedefs.

All non-corex (NVIDIA CUDA) code paths are guarded by
UIPC_MUDA_USE_COREX and remain bit-identical to upstream.

Made-with: Cursor

+++ src/backends/cuda/algorithm/corex_matrix_converter_kernels.h
@@ -3,6 +3,7 @@
 
 #include <cstdint>
 #include <Eigen/Core>
+#include <uipc/common/type_define.h>
 
 namespace uipc::backend::cuda::corex_matconv
 {
@@ -22,7 +23,7 @@ void launch_write_unique_indices(int N, const int* unique_indices, int* dst_indi
 void launch_scatter_col_counts(int N, const int* unique_indices,
                                const int* counts, int* col_counts_per_row);
 
-using BlockT3 = Eigen::Matrix<double, 3, 3>;
+using BlockT3 = Eigen::Matrix<Float, 3, 3>;
 
 void launch_copy_sorted_blocks_3x3(int N, const BlockT3* src_blocks,
                                     const int* sort_index, BlockT3* dst_blocks);
@@ -37,7 +38,7 @@ void launch_segmental_reduce_3x3(int N, const int* segment_ids,
                                   const BlockT3* in_blocks, BlockT3* out_blocks,
                                   int out_count);
 
-using VecT3 = Eigen::Matrix<double, 3, 1>;
+using VecT3 = Eigen::Matrix<Float, 3, 1>;
 
 void launch_segmental_reduce_3x1(int N, const int* segment_ids,
                                   const VecT3* in_vecs, VecT3* out_vecs,
